pyproject.toml
[project]
name = "discordmoviebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "google-api-python-client>=2.151.0",
    "google-auth>=2.36.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "gspread>=6.1.4",
]
